Refactor: move classes to own files
