+++ src/mailmerge/mailmerge.py
@@ -1,7 +1,5 @@
-import datetime
 import os
 import re
-import shlex
 import warnings
 from copy import deepcopy
 
@@ -10,48 +8,11 @@ from zipfile import ZIP_DEFLATED, ZipFile
 
 from lxml import etree
 
-NAMESPACES = {
-    "w": "http://schemas.openxmlformats.org/wordprocessingml/2006/main",
-    "wp": "http://schemas.openxmlformats.org/drawingml/2006/wordprocessingDrawing",
-    "mc": "http://schemas.openxmlformats.org/markup-compatibility/2006",
-    "ct": "http://schemas.openxmlformats.org/package/2006/content-types",
-    "rr": "http://schemas.openxmlformats.org/package/2006/relationships",
-    "r": "http://schemas.openxmlformats.org/officeDocument/2006/relationships",
-    "xml": "http://www.w3.org/XML/1998/namespace",
-}
-
-ATTACHMENT_TAGS = [
-    "hdr",  # header
-    "ftr",  # footer
-    "footnotes",  # footnotes
-    "endnotes",  # endnotes
-]
-
-ATTACHMENT_TAGS_WITH_NAMESPACE = {"{%(w)s}" % NAMESPACES + tag for tag in ATTACHMENT_TAGS}
-
-CONTENT_TYPES_PARTS = {
-    "application/vnd.openxmlformats-officedocument.wordprocessingml.document.main+xml": "main",
-    "application/vnd.ms-word.document.macroEnabled.main+xml": "main",
-    "application/vnd.openxmlformats-officedocument.wordprocessingml.header+xml": "header_footer",
-    "application/vnd.openxmlformats-officedocument.wordprocessingml.footer+xml": "header_footer",
-    "application/vnd.openxmlformats-officedocument.wordprocessingml.footnotes+xml": "notes",
-    "application/vnd.openxmlformats-officedocument.wordprocessingml.endnotes+xml": "notes",
-    "application/vnd.openxmlformats-officedocument.wordprocessingml.settings+xml": "settings",
-}
-
-VALID_SEPARATORS = {
-    "page_break",
-    "column_break",
-    "textWrapping_break",
-    "continuous_section",
-    "evenPage_section",
-    "nextColumn_section",
-    "nextPage_section",
-    "oddPage_section",
-}
+from .constants import CONTENT_TYPES_PARTS, MAKE_TESTS_HAPPY, NAMESPACES, VALID_SEPARATORS
+from .field import SimpleMergeField, SkipRecord
+from .mergedata import MergeData
 
 PARTFILENAME_RE = re.compile(r"([A-Za-z_]+)(\d+).xml")
-IDSTR_RE = re.compile(r"([A-Za-z_]+)(\d+)")
 NUMBERFORMAT_RE = re.compile(r"([^0.,'#PN]+)?(P\d+|N\d+|[0.,'#]+%?)([^0.,'#%].*)?")
 DATEFORMAT_RE = "|".join([r"{}+".format(switch) for switch in "yYmMdDhHsS"] + [r"am/pm", r"AM/PM"])
 DATEFORMAT_MAP = {
@@ -82,538 +43,6 @@ DATEFORMAT_MAP = {
     "am/pm": "%p",
     "AM/PM": "%p",
 }
-
-TAGS_WITH_ID = {"wp:docPr": {"name": "Picture {id}"}}
-
-MAKE_TESTS_HAPPY = True
-
-
-class NextRecord(Exception):
-    pass
-
-
-class SkipRecord(Exception):
-    pass
-
-
-class MergeField(object):
-    """
-    Base MergeField class
-
-    it contains the field name,
-    and a method to fill a list of elements (runs) given the data dictionary
-    That filled_elements list should be inserted in the document
-    """
-
-    def __init__(
-        self,
-        parent,
-        name="",
-        key=None,
-        instr=None,
-        instr_tokens=None,
-        nested=False,
-        all_elements=None,
-        instr_elements=None,
-        show_elements=None,
-    ):
-        """Inits the MergeField class
-
-        Args:
-            parent: The parent element of the MergeField in the tree.
-            idx: The idx of the MergeField in the parent.
-            name: The name of the field, if applicable
-            all_elements: The list of all elements that need to be replaced in the parent
-            instr_elements: Elements that show the instructions of the field (used to construct the value)
-            show_elements: Elements that show the current value of the field
-        """
-        self.parent = parent
-        self.nested = nested
-        # the key of this MergeField to be able to identify it. It is used as the name in
-        # the replaced MergeField element
-        self.key = key
-        # the list of elements to add when merging
-        self._all_elements = [] if all_elements is None else all_elements
-        self._instr_elements = [] if instr_elements is None else instr_elements
-        self._show_elements = [] if show_elements is None else show_elements
-        self.instr = instr
-        instr_tokens = [] if instr_tokens is None else instr_tokens
-        self.instr_tokens = instr_tokens
-        self.filled_elements = []
-        self.filled_value = None
-        self.name = name
-        if not name and instr_tokens[1:]:
-            self.name = instr_tokens[1]
-
-    def reset(self):
-        """resets the value"""
-        self.filled_elements = []
-
-    def _format(self, value):
-        options = self.instr_tokens[2:]
-        while options:
-            flag = options[0][0:2]
-            if not flag:
-                options = options[1:]
-                continue
-            if options[0][2:]:  # no space after the flag
-                option = options[0][2:]
-                options = options[1:]
-            else:
-                option = options[1]
-                options = options[2:]
-            if flag in ("\\b", "\\f"):
-                value = self._format_bf(value, flag, option)
-            if flag in ("\\#"):
-                value = self._format_number(value, flag, option)
-            if flag in ("\\@"):
-                value = self._format_date(value, flag, option)
-            if flag in ("\\*"):
-                value = self._format_text(value, flag, option)
-
-        if isinstance(value, (datetime.datetime, datetime.date, datetime.time)):
-            # TODO format the date according to the locale -- set the locale
-            date_formats = []
-            if hasattr(value, "month"):
-                date_formats.append("%x")
-            if hasattr(value, "hour"):
-                date_formats.append("%X")
-            value = value.strftime(" ".join(date_formats))
-
-        return value
-
-    def _format_bf(self, value, flag, option):
-        # print("<{}><{}>".format(value, type(value)))
-        if value:
-            if flag == "\\b":
-                return option + str(value)
-            return str(value) + option
-        # no value no text
-        return None
-
-    def _format_text(self, value, flag, option):
-        option = option.lower()
-        if option == "caps":
-            return str(value).title()
-        if option == "firstcap":
-            return str(value).capitalize()
-        if option == "upper":
-            return str(value).upper()
-        if option == "lower":
-            return str(value).lower()
-
-        return value
-
-    def _format_number(self, value, flag, option):
-        format_match = NUMBERFORMAT_RE.match(option)
-        if value is None:
-            value = 0
-        if format_match is None:
-            warnings.warn("Non conforming number format <{}>".format(option))
-            return value
-        format_prefix = format_match.group(1) or ""
-        format_number = format_match.group(2)
-        format_suffix = format_match.group(3) or ""
-        if format_number[0] == "P":
-            return "{{}}{{:.{}%}}{{}}".format(int(format_number[1:])).format(format_prefix, value, format_suffix)
-        if format_number[0] == "N":
-            return "{{}}{{:.{}f}}{{}}".format(int(format_number[1:])).format(format_prefix, value, format_suffix)
-        if format_number[-1] == "%":
-            return "{}{:.0%}{}".format(format_prefix, value, format_suffix)
-        thousand_info = [("_", thousand_char) for thousand_char in "'," if thousand_char in format_number] + [("", "")]
-        thousand_flag, thousand_char = thousand_info[0]
-        format_number = format_number.replace(",", "")
-        digits, decimals = (format_number.split(".") + [""])[0:2]
-        zero_digits = len(digits.replace("#", ""))
-        _zero_decimals = len(decimals.replace("#", ""))
-        len_decimals_plus_dot = 0 if not decimals else 1 + len(decimals)
-        number_format_text = "{{}}{{:{zero_digits}{thousand_flag}{decimals}f}}{{}}".format(
-            thousand_flag=thousand_flag,
-            zero_digits="0>{}".format(zero_digits + len_decimals_plus_dot) if zero_digits > 1 else "",
-            decimals=".{}".format(len(decimals)),
-        )
-        # print(self.name, "<", option, ">", number_format_text)
-        try:
-            result = number_format_text.format(format_prefix, value, format_suffix)
-            if thousand_flag:
-                result = result.replace(thousand_flag, thousand_char)
-            return result
-        except Exception as e:
-            raise ValueError("Invalid number format <{}> with error <{}>".format(number_format_text, e))
-
-    def _format_date(self, value, flag, option):
-        if value is None:
-            return ""
-
-        if not isinstance(value, (datetime.datetime, datetime.date, datetime.time)):
-            return str(value)
-
-        # set the locale to be used for time
-        # more checking needed before activating
-        # locale.setlocale(locale.LC_TIME, "")
-        fmt = re.sub(DATEFORMAT_RE, lambda x: DATEFORMAT_MAP[x[0]], option)
-        fmt_args = {"d": value, "hour12": 0}
-        if isinstance(value, (datetime.datetime, datetime.time)):
-            fmt_args["hour12"] = value.hour % 12
-        fmt = fmt.format(**fmt_args)
-        value = value.strftime(fmt)
-        # warnings.warn("Date formatting not yet implemented <{}>".format(option))
-        return value
-
-    def fill_data(self, merge_data, row):
-        """fills the filled_elements with all the elements containing the output text"""
-        self.filled_elements = []
-        value = row.get(self.name, "UNKNOWN({})".format(self.instr))
-        try:
-            value = self._format(value)
-        except Exception as e:
-            warnings.warn("Invalid formatting for field <{}> with error <{}>".format(self.instr, e))
-            # raise
-
-        self.filled_value = value
-
-        if value is None:
-            # no elements should be filled ?
-            # or empty text ?
-            if MAKE_TESTS_HAPPY:
-                value = ""
-            else:
-                return
-
-        elem = deepcopy(self._instr_elements[0])
-        for child in elem.xpath("w:instrText", namespaces=NAMESPACES):
-            elem.remove(child)
-        for child in elem.xpath("w:t", namespaces=NAMESPACES):
-            elem.remove(child)
-
-        text_parts = str(value).replace("\r", "").split("\n")
-        elem.append(self._make_text(text_parts[0]))
-        for text_part in text_parts[1:]:
-            elem.append(self._make_br())
-            elem.append(self._make_text(text_part))
-
-        self.filled_elements.append(elem)
-
-    def get_elements_to_replace(self, keep_field=False):
-        """returns the list of filled elements to put in the document
-
-        three possible outcomes:
-        - only the value (keep_field=False)
-        - the original field without updating the value(keep_field=True, no value filled in)
-        - the field with an updated value(keep_field=True, new value filled in)
-        """
-        if keep_field:
-            if not self.filled_elements:  # we keep the original value
-                return self._all_elements
-            return self.get_field_with_filled_elements()
-        return self.filled_elements
-
-    def get_field_with_filled_elements(self):
-        # for complex fields
-        all_elements = self._all_elements[:]  # copy of all elements
-        if not self._show_elements:
-            separate_element = deepcopy(self._all_elements[-1])
-            separate_element.find("w:fldChar", namespaces=NAMESPACES).set("{%(w)s}fldCharType" % NAMESPACES, "separate")
-            all_elements[-1:-1] = [separate_element] + self.filled_elements
-        else:
-            index = all_elements.index(self._show_elements[0])
-            all_elements[index : index + len(self._show_elements)] = self.filled_elements
-        return all_elements
-
-    def _make_br(self):
-        return etree.Element("{%(w)s}br" % NAMESPACES, attrib=None, nsmap=None)
-
-    def _make_text(self, text):
-        if self.nested:
-            text_node = etree.Element("{%(w)s}instrText" % NAMESPACES, attrib=None, nsmap=None)
-            text_node.set("{%(xml)s}space" % NAMESPACES, "preserve")
-        else:
-            text_node = etree.Element("{%(w)s}t" % NAMESPACES, attrib=None, nsmap=None)
-
-        text_node.text = text
-        return text_node
-
-    def insert_into_tree(self):
-        """inserts a MergeField element in the original tree at the right position"""
-        # Make sure ALL elements from the original tree are removed except for the first useful, that we will replace
-        for subelem in self._all_elements[1:]:
-            self.parent.remove(subelem)
-
-        replacement_element = etree.Element("MergeField", attrib=None, nsmap=None, merge_key=self.key, name=self.name)
-        self.parent.replace(self._all_elements[0], replacement_element)
-        return replacement_element
-
-
-class SimpleMergeField(MergeField):
-    """differences for simple fields"""
-
-    def get_field_with_filled_elements(self):
-        # for simple fields
-        field_element = deepcopy(self._all_elements[0])
-        # remove all child elements
-        for child in list(field_element):
-            field_element.remove(child)
-        for child in self.filled_elements:
-            field_element.append(child)
-        return [field_element]
-
-
-class NextField(MergeField):
-    def fill_data(self, merge_data, row):
-        raise NextRecord()
-
-
-class UniqueIdsManager(object):
-    """handles different counters for various ids in the document"""
-
-    def __init__(self):
-        self.id_type_map = {}  # type of id -> {'max': max_id, 'ids': set(existing_ids)}
-
-    def register_id(self, id_type, obj_id=None):
-        """registers an new object id or creates a new id for the type"""
-        type_id_value = self.id_type_map.setdefault(id_type, {"max": 0, "ids": set()})
-        new_obj_id = None
-        if obj_id is None or obj_id in type_id_value["ids"]:
-            obj_id = type_id_value["max"] + 1
-            new_obj_id = obj_id
-        type_id_value["ids"].add(obj_id)
-        type_id_value["max"] = max(type_id_value["max"], obj_id)
-        # print("registered", id_type, obj_id, new_obj_id, "max", type_id_value['max'])
-        return new_obj_id
-
-    def register_id_str(self, id_str):
-        """registers directly a string of format 'type1231' where the id_type is before the id"""
-        match = IDSTR_RE.match(id_str)
-        assert match
-        id_type, obj_id = match.groups()
-        new_obj_id = self.register_id(id_type, obj_id=int(obj_id))
-        if new_obj_id is not None:
-            # print(id_type, obj_id, new_obj_id, self.id_type_map[id_type])
-            return "%s%d" % (id_type, new_obj_id)
-
-
-class MergeData(object):
-    """prepare the MergeField objects and the data"""
-
-    SUPPORTED_FIELDS = {"MERGEFIELD", "NEXT"}
-    FIELD_CLASSES = {"NEXT": NextField}
-
-    def __init__(self, remove_empty_tables=False, keep_fields="none"):
-        self._merge_field_map = {}  # merge_field.key: MergeField()
-        self._merge_field_next_id = 0
-        self.unique_id_manager = UniqueIdsManager()
-        self.has_nested_fields = False
-        self.remove_empty_tables = remove_empty_tables
-        self.keep_fields = keep_fields
-        self.replace_fields_with_missing_data = False
-        self._rows = None
-        self._current_index = None
-
-    def start_merge(self, replacements):
-        assert self._rows is None, "merge already started"
-        self._rows = replacements
-        return self.next_row()
-
-    def next_row(self):
-        assert self._rows is not None, "merge not yet started"
-
-        if self._current_index is None:
-            self._current_index = 0
-        else:
-            self._current_index += 1
-
-        if self._current_index < len(self._rows):
-            return self._rows[self._current_index]
-
-    def is_first(self):
-        return self._current_index == 0
-
-    def get_new_element_id(self, element):
-        """Returns None if the existing id is new otherwise a new id"""
-        # tag = element.tag
-        elem_id = element.get("id")
-        if elem_id is None:
-            return None
-        elem_id = int(elem_id)
-        new_id = self.unique_id_manager.register_id("id", elem_id)
-        if new_id:
-            return str(new_id)
-        return None
-
-    def get_merge_fields(self, key):
-        merge_obj = self.get_field_obj(key)
-        if merge_obj.name:
-            yield merge_obj.name
-
-    def get_instr_text(self, elements, recursive=False):
-        return "".join(
-            [
-                text
-                for elem in elements
-                for text in elem.xpath("w:instrText/text()", namespaces=NAMESPACES)
-                + [
-                    "{{{}}}".format(obj_name) if not recursive else self.get_field_obj(obj_name).instr
-                    for obj_name in elem.xpath("@merge_key")
-                ]
-            ]
-        )
-
-    @classmethod
-    def _get_instr_tokens(cls, instr):
-        s = shlex.shlex(instr, posix=True)
-        s.whitespace_split = True
-        s.commenters = ""
-        s.escape = ""
-        return s
-
-    @classmethod
-    def _get_field_type(cls, instr):
-        s = shlex.split(instr, posix=False)
-        # For elements with no instr, otherwise this will throw an exception
-        if len(s) > 0:
-            return s[0], s[1:]
-        return "", []
-
-    def make_data_field(
-        self,
-        parent,
-        field_class=MergeField,
-        key=None,
-        nested=False,
-        instr=None,
-        all_elements=None,
-        instr_elements=None,
-        show_elements=None,
-        **kwargs,
-    ):
-        """MergeField factory method"""
-        if key is None:
-            key = self._get_next_key()
-
-        instr = instr or self.get_instr_text(instr_elements)
-        field_type, rest = self._get_field_type(instr)
-        if field_type not in self.SUPPORTED_FIELDS:
-            # ignore the field
-            # print("ignore field", instr)
-            return None
-        field_class = self.FIELD_CLASSES.get(field_type, field_class)
-
-        try:
-            tokens = list(self._get_instr_tokens(instr))
-        except ValueError as e:
-            tokens = [field_type] + list(map(lambda part: part.replace('"', ""), rest))
-            warnings.warn("Invalid field description <{}> near: <{}>".format(str(e), instr))
-
-        # print("make data object", field_class, instr, len(elements), len(kwargs.get('ignore_elements', [])))
-        field_obj = field_class(
-            parent,
-            key=key,
-            instr=instr,
-            nested=nested,
-            instr_tokens=tokens,
-            all_elements=all_elements,
-            instr_elements=instr_elements,
-            show_elements=show_elements,
-            **kwargs,
-        )
-
-        assert key not in self._merge_field_map
-        if nested:
-            self.has_nested_fields = True
-        self._merge_field_map[key] = field_obj
-        return field_obj
-
-    def get_field_obj(self, key):
-        return self._merge_field_map[key]
-
-    def mark_field_as_nested(self, key, nested=True):
-        if nested:
-            self.has_nested_fields = True
-        self.get_field_obj(key).nested = nested
-
-    def _get_next_key(self):
-        key = "field_{}".format(self._merge_field_next_id)
-        self._merge_field_next_id += 1
-        return key
-
-    def replace(self, body, row):
-        """replaces in the body xml tree the MergeField elements with values from the row"""
-        all_tables = {key: value for key, value in row.items() if isinstance(value, list)}
-
-        for anchor, table_rows in all_tables.items():
-            self.replace_table_rows(body, anchor, table_rows)
-
-        merge_fields = body.findall(".//MergeField")
-        for field_element in merge_fields:
-            field_obj = None
-            try:
-                field_obj = self.get_field_object(field_element, row)
-                field_obj.reset()
-                if self._has_value_in_row(field_element, row):
-                    field_obj.fill_data(self, row)  # can throw NextRecord
-                    self.replace_field(field_element, field_obj)
-                elif self.replace_fields_with_missing_data:
-                    self.replace_field(field_element, field_obj, force_keep_field=True)
-            except NextRecord:
-                self.replace_field(field_element, field_obj)
-                row = self.next_row()
-
-    def _has_value_in_row(self, field_element, row):
-        return not (field_element.get("name") and (row is None or field_element.get("name") not in row))
-
-    def replace_field(self, field_element, field_obj=None, force_keep_field=False):
-        """replaces a field element MergeField in the body with the filled_elements"""
-        # assert len(filled_field.filled_elements) == 1
-        if field_obj:
-            keep_field = force_keep_field or self.keep_fields == "all"
-            elements_to_replace = field_obj.get_elements_to_replace(keep_field=keep_field)
-            for text_element in reversed(elements_to_replace):
-                field_element.addnext(text_element)
-        field_element.getparent().remove(field_element)
-
-    def replace_table_rows(self, body, anchor, rows):
-        """replace the rows of a table with the values from the rows list"""
-        for table, idx, template in self.__find_row_anchor(body, anchor):
-            if len(rows) > 0:
-                del table[idx]
-                for i, row_data in enumerate(rows):
-                    row = deepcopy(template)
-                    self.replace(row, row_data)
-                    table.insert(idx + i, row)
-            else:
-                # if there is no data for a given table
-                # we check whether table needs to be removed
-                if self.remove_empty_tables:
-                    parent = table.getparent()
-                    parent.remove(table)
-
-    def __find_row_anchor(self, body, field):
-        for table in body.findall(".//{%(w)s}tbl" % NAMESPACES):
-            for idx, row in enumerate(table):
-                if row.find('.//MergeField[@name="%s"]' % field) is not None:
-                    yield table, idx, row
-
-    def get_field_object(self, field_element, row):
-        """ " fills the corresponding MergeField python object with data from row"""
-        # if field_element.get('name') and (row is None or field_element.get('name') not in row):
-        #     return None
-        field_key = field_element.get("merge_key")
-        field_obj = self._merge_field_map[field_key]
-        return field_obj
-
-    def fix_id(self, element, attr_gen):
-        """will replace an id with a new unique id"""
-        new_id = self.get_new_element_id(element)
-        if new_id is not None:
-            element.attrib["id"] = new_id
-            for attr_name, attr_value in attr_gen.items():
-                element.attrib[attr_name] = attr_value.format(id=new_id)
-
-    def fix_ids(self, current_body):
-        """will fix all ids in the current body"""
-        for tag, attr_gen in TAGS_WITH_ID.items():
-            for elem in current_body.xpath("//{}".format(tag), namespaces=NAMESPACES):
-                self.fix_id(elem, attr_gen)
 
 
 class RelationsDocument(object):
